[REFACTOR/#111] marketProduct 서비스 분리
