+++ app/src/main/java/com/napzak/market/data/marketinfo/service/MarketInfoService.kt
@@ -1,12 +1,9 @@
 package com.napzak.market.data.marketinfo.service
 
 import com.napzak.market.core.network.BaseResponse
-import com.napzak.market.data.marketinfo.dto.MarketProductBuyItemsResponse
-import com.napzak.market.data.marketinfo.dto.MarketProductSellItemsResponse
 import com.napzak.market.data.marketinfo.dto.MarketInfoResponse
 import retrofit2.http.GET
 import retrofit2.http.Path
-import retrofit2.http.Query
 
 interface MarketInfoService {
     @GET("stores/{storeId}")
@@ -14,36 +11,4 @@ interface MarketInfoService {
         @Path("storeId")
         storeId: Long,
     ): BaseResponse<MarketInfoResponse>
-
-    @GET("products/sell/stores/{storeOwnerId}")
-    suspend fun getMarketProductSellItems(
-        @Path("storeOwnerId")
-        storeId: Long,
-        @Query("sortOption")
-        sortOption: String = "RECENT",
-        @Query("isOnSale")
-        isOnSale: Boolean = false,
-        @Query("isUnopened")
-        isUnopened: Boolean = false,
-        @Query("genreId")
-        genreIds: List<Long>? = null,
-        @Query("cursor")
-        cursor: String = "",
-    ): BaseResponse<MarketProductSellItemsResponse>
-
-    @GET("products/buy/stores/{storeOwnerId}")
-    suspend fun getMarketProductBuyItems(
-        @Path("storeOwnerId")
-        storeId: Long,
-        @Query("sortOption")
-        sortOption: String = "RECENT",
-        @Query("isOnSale")
-        isOnSale: Boolean = false,
-        @Query("isUnopened")
-        isUnopened: Boolean = false,
-        @Query("genreId")
-        genreIds: List<Long>? = null,
-        @Query("cursor")
-        cursor: String = "",
-    ): BaseResponse<MarketProductBuyItemsResponse>
 }
